refector(147366): Muda as funções de filtrar e obter classificações de dietas no recreio para outro arquivo
